+++ src/commands/slash.js
@@ -32,6 +32,7 @@ import { reverse, reverseTranslator } from "./reverse.js";
 import saveLog from "./save-log.js";
 import shuffle from "./shuffle.js";
 import spotify from "./spotify.js";
+import stats from "./stats.js";
 
 import { interactionReply, isReleasedCommand, isSentinelle } from "./utils.js";
 
@@ -194,6 +195,7 @@ const slashCommands = [
   roll,
   shuffle,
   spotify,
+  stats,
 ]; //command + action
 
 // HELP