+++ changelog.md
@@ -1,4 +1,11 @@
-﻿## 5.4.0
+﻿## 5.5.0
+- Feat: statistics ([#234](https://github.com/Eccleria/ewibot/pull/234))
+	- add `/stats use` command to dis/enable Ewibot stats for requesting user
+	- count `gifs`, `hello` words, `hungry` words, and `emojis` for accepting users
+	- count `emotes` occurences and `cats` pictures sent for whole guild
+- Feat: `dbReturnType` frozen object with returned values for db helpers ([#234](https://github.com/Eccleria/ewibot/pull/234))
+
+## 5.4.0
 - Feat: add reaction if is mentioned by a user ([#230](https://github.com/Eccleria/ewibot/pull/230))
 - Feat: add `DEBUGLOGS` .env attribute - this add console.logs when switched to "yes"
 - Fix: wrong `apology` detection